+++ src/imio/history/tests/test_contenthistory.py
@@ -150,7 +150,17 @@ class TestContentHistory(IntegrationTestCase):
         # render comments will first try to translate the comments
         # then turn it from 'text/plain' to 'text/html'
         view = getMultiAdapter((self.doc, self.portal.REQUEST), name='contenthistory')
-        self.assertEquals(view.renderComments('data_change'),
-                          u'<p>Data change</p>')
-        self.assertEquals(view.renderComments('Custom comments not translatable.\nAnd one additional line.'),
+        view.histories_to_handle = (u'workflow', )
+
+        # translated
+        # do a transition with a comment that will be translatable
+        self.wft.doActionFor(self.doc, 'publish', comment='data_change')
+        last_event = view.getHistory()[0]
+        self.assertEquals(view.renderComments(last_event), u'<p>Data change</p>')
+
+        # turned from text/plain to text/html
+        # do a transition with a comment that will be translatable
+        self.wft.doActionFor(self.doc, 'retract', comment='Custom comments not translatable.\nAnd one additional line.')
+        last_event = view.getHistory()[0]
+        self.assertEquals(view.renderComments(last_event),
                           u'<p>Custom comments not translatable.<br />And one additional line.</p>')